Moved most references to AWS into separate file
