+++ plugins/aws/src/s3Database.ts
@@ -1,21 +1,15 @@
-import {
-  CloudFrontClient,
-  CreateInvalidationCommand,
-} from "@aws-sdk/client-cloudfront";
-import {
-  DeleteObjectCommand,
-  GetObjectCommand,
-  ListObjectsV2Command,
-  NoSuchKey,
-  S3Client,
-  type S3ClientConfig,
-} from "@aws-sdk/client-s3";
-import { Upload } from "@aws-sdk/lib-storage";
+import { CloudFrontClient } from "@aws-sdk/client-cloudfront";
+import { S3Client, type S3ClientConfig } from "@aws-sdk/client-s3";
 import type { Bundle, DatabasePluginHooks } from "@hot-updater/plugin-core";
 import { createDatabasePlugin } from "@hot-updater/plugin-core";
 import { orderBy } from "es-toolkit";
-import mime from "mime";
-import { streamToString } from "./utils/streamToString";
+import {
+  deleteObject,
+  invalidateCloudFront,
+  listObjects,
+  loadJsonFromS3,
+  uploadJsonToS3,
+} from "./awsHooks";
 
 export interface S3DatabaseConfig extends S3ClientConfig {
   bucketName: string;
@@ -25,79 +19,6 @@ export interface S3DatabaseConfig extends S3ClientConfig {
 interface BundleWithUpdateJsonKey extends Bundle {
   _updateJsonKey: string;
   _oldUpdateJsonKey?: string;
-}
-
-/**
- * Loads JSON data from S3.
- * Returns null if NoSuchKey error occurs.
- */
-async function loadJsonFromS3<T>(
-  client: S3Client,
-  bucket: string,
-  key: string,
-): Promise<T | null> {
-  try {
-    const { Body } = await client.send(
-      new GetObjectCommand({ Bucket: bucket, Key: key }),
-    );
-    if (!Body) return null;
-    const bodyContents = await streamToString(Body);
-    return JSON.parse(bodyContents) as T;
-  } catch (e) {
-    if (e instanceof NoSuchKey) return null;
-    throw e;
-  }
-}
-
-/**
- * Converts data to JSON string and uploads to S3.
- */
-async function uploadJsonToS3<T>(
-  client: S3Client,
-  bucket: string,
-  key: string,
-  data: T,
-) {
-  const Body = JSON.stringify(data);
-  const ContentType = mime.getType(key) ?? "application/json";
-  const upload = new Upload({
-    client,
-    params: {
-      Bucket: bucket,
-      Key: key,
-      Body,
-      ContentType,
-      CacheControl: "max-age=31536000",
-    },
-  });
-  await upload.done();
-}
-
-/**
- * Invalidates CloudFront cache for the given paths.
- */
-async function invalidateCloudFront(
-  client: CloudFrontClient,
-  distributionId: string,
-  paths: string[],
-) {
-  if (paths.length === 0) {
-    return;
-  }
-
-  const timestamp = new Date().getTime();
-  await client.send(
-    new CreateInvalidationCommand({
-      DistributionId: distributionId,
-      InvalidationBatch: {
-        CallerReference: `invalidation-${timestamp}`,
-        Paths: {
-          Quantity: paths.length,
-          Items: paths,
-        },
-      },
-    }),
-  );
 }
 
 // Helper function to remove internal management keys
@@ -112,14 +33,12 @@ function removeBundleInternalKeys(bundle: BundleWithUpdateJsonKey): Bundle {
  * - If a channel is provided, only that channel's update.json files are listed.
  * - Otherwise, all channels for the given platform are returned.
  */
-async function listUpdateJsonKeys(
+export async function listUpdateJsonKeys(
   client: S3Client,
   bucketName: string,
   platform?: string,
-  channel?: string,
+  channel?: string
 ): Promise<string[]> {
-  let continuationToken: string | undefined;
-  const keys: string[] = [];
   const prefix = channel
     ? platform
       ? `${channel}/${platform}/`
@@ -131,23 +50,11 @@ async function listUpdateJsonKeys(
       ? new RegExp(`^${channel}/${platform}/[^/]+/update\\.json$`)
       : new RegExp(`^${channel}/[^/]+/[^/]+/update\\.json$`)
     : platform
-      ? new RegExp(`^[^/]+/${platform}/[^/]+/update\\.json$`)
-      : /^[^\/]+\/[^\/]+\/[^\/]+\/update\.json$/;
-  do {
-    const response = await client.send(
-      new ListObjectsV2Command({
-        Bucket: bucketName,
-        Prefix: prefix,
-        ContinuationToken: continuationToken,
-      }),
-    );
-    const found = (response.Contents ?? [])
-      .map((item) => item.Key)
-      .filter((key): key is string => !!key && pattern.test(key));
-    keys.push(...found);
-    continuationToken = response.NextContinuationToken;
-  } while (continuationToken);
-  return keys;
+    ? new RegExp(`^[^/]+/${platform}/[^/]+/update\\.json$`)
+    : /^[^\/]+\/[^\/]+\/[^\/]+\/update\.json$/;
+
+  const keys = await listObjects(client, bucketName);
+  return keys.filter((key): key is string => !!key && pattern.test(key));
 }
 
 /**
@@ -157,38 +64,22 @@ async function listUpdateJsonKeys(
 async function updateTargetVersionsForPlatform(
   client: S3Client,
   bucketName: string,
-  platform: string,
+  platform: string
 ): Promise<Set<string>> {
   // Retrieve all update.json files for the platform across channels.
-  let continuationToken: string | undefined;
-  const keys: string[] = [];
   const pattern = new RegExp(`^[^/]+/${platform}/[^/]+/update\\.json$`);
-  do {
-    const response = await client.send(
-      new ListObjectsV2Command({
-        Bucket: bucketName,
-        Prefix: "",
-        ContinuationToken: continuationToken,
-      }),
-    );
-    const found = (response.Contents ?? [])
-      .map((item) => item.Key)
-      .filter((key): key is string => !!key && pattern.test(key));
-    keys.push(...found);
-    continuationToken = response.NextContinuationToken;
-  } while (continuationToken);
+  const keys = (await listUpdateJsonKeys(client, bucketName, platform)).filter(
+    (key) => pattern.test(key)
+  );
 
   // Group keys by channel (channel is the first part of the key)
-  const keysByChannel = keys.reduce(
-    (acc, key) => {
-      const parts = key.split("/");
-      const channel = parts[0];
-      acc[channel] = acc[channel] || [];
-      acc[channel].push(key);
-      return acc;
-    },
-    {} as Record<string, string[]>,
-  );
+  const keysByChannel = keys.reduce((acc, key) => {
+    const parts = key.split("/");
+    const channel = parts[0];
+    acc[channel] = acc[channel] || [];
+    acc[channel].push(key);
+    return acc;
+  }, {} as Record<string, string[]>);
 
   const updatedTargetFiles = new Set<string>();
 
@@ -200,7 +91,7 @@ async function updateTargetVersionsForPlatform(
     const oldTargetVersions =
       (await loadJsonFromS3<string[]>(client, bucketName, targetKey)) ?? [];
     const newTargetVersions = oldTargetVersions.filter((v) =>
-      currentVersions.includes(v),
+      currentVersions.includes(v)
     );
     for (const v of currentVersions) {
       if (!newTargetVersions.includes(v)) newTargetVersions.push(v);
@@ -219,7 +110,7 @@ async function updateTargetVersionsForPlatform(
 
 export const s3Database = (
   config: S3DatabaseConfig,
-  hooks?: DatabasePluginHooks,
+  hooks?: DatabasePluginHooks
 ) => {
   const { bucketName, cloudfrontDistributionId, ...s3Config } = config;
   const client = new S3Client(s3Config);
@@ -241,7 +132,7 @@ export const s3Database = (
 
     const platformPromises = PLATFORMS.map(async (platform) => {
       // Retrieve update.json files for the platform across all channels.
-      const keys = await listUpdateJsonKeys(client, bucketName, platform);
+      const keys = await listObjects(client, bucketName);
       const filePromises = keys.map(async (key) => {
         const bundlesData =
           (await loadJsonFromS3<Bundle[]>(client, bucketName, key)) ?? [];
@@ -297,7 +188,7 @@ export const s3Database = (
               ([key, value]) =>
                 value === undefined ||
                 value === null ||
-                bundle[key as keyof Bundle] === value,
+                bundle[key as keyof Bundle] === value
             );
           });
         }
@@ -338,7 +229,7 @@ export const s3Database = (
 
             changedBundlesByKey[key] = changedBundlesByKey[key] || [];
             changedBundlesByKey[key].push(
-              removeBundleInternalKeys(bundleWithKey),
+              removeBundleInternalKeys(bundleWithKey)
             );
 
             // CloudFront 무효화를 위한 경로 추가
@@ -383,7 +274,7 @@ export const s3Database = (
               pendingBundlesMap.set(data.id, updatedBundle);
 
               changedBundlesByKey[newKey].push(
-                removeBundleInternalKeys(updatedBundle),
+                removeBundleInternalKeys(updatedBundle)
               );
 
               // Add paths for CloudFront invalidation
@@ -400,7 +291,7 @@ export const s3Database = (
             changedBundlesByKey[currentKey] =
               changedBundlesByKey[currentKey] || [];
             changedBundlesByKey[currentKey].push(
-              removeBundleInternalKeys(updatedBundle),
+              removeBundleInternalKeys(updatedBundle)
             );
 
             // CloudFront 무효화를 위한 경로 추가
@@ -415,13 +306,11 @@ export const s3Database = (
               (await loadJsonFromS3<Bundle[]>(client, bucketName, oldKey)) ??
               [];
             const updatedBundles = currentBundles.filter(
-              (b) => !removalsByKey[oldKey].includes(b.id),
+              (b) => !removalsByKey[oldKey].includes(b.id)
             );
             updatedBundles.sort((a, b) => b.id.localeCompare(a.id));
             if (updatedBundles.length === 0) {
-              await client.send(
-                new DeleteObjectCommand({ Bucket: bucketName, Key: oldKey }),
-              );
+              await deleteObject(client, bucketName, oldKey);
             } else {
               await uploadJsonToS3(client, bucketName, oldKey, updatedBundles);
             }
@@ -434,11 +323,11 @@ export const s3Database = (
             const currentBundles =
               (await loadJsonFromS3<Bundle[]>(client, bucketName, key)) ?? [];
             const pureBundles = changedBundlesByKey[key].map(
-              (bundle) => bundle,
+              (bundle) => bundle
             );
             for (const changedBundle of pureBundles) {
               const index = currentBundles.findIndex(
-                (b) => b.id === changedBundle.id,
+                (b) => b.id === changedBundle.id
               );
               if (index >= 0) {
                 currentBundles[index] = changedBundle;
@@ -457,7 +346,7 @@ export const s3Database = (
           const updatedPaths = await updateTargetVersionsForPlatform(
             client,
             bucketName,
-            platform,
+            platform
           );
           for (const path of updatedPaths) {
             updatedTargetFilePaths.add(path);
@@ -478,7 +367,7 @@ export const s3Database = (
           await invalidateCloudFront(
             cloudfrontClient,
             cloudfrontDistributionId,
-            Array.from(pathsToInvalidate),
+            Array.from(pathsToInvalidate)
           );
         }
 
@@ -486,6 +375,6 @@ export const s3Database = (
         hooks?.onDatabaseUpdated?.();
       },
     },
-    hooks,
+    hooks
   );
 };
